Added URL form prefills to Contact page
The contact page now allows you to prefill fields by id. To do so, place
a hash at the end of the URL and put the field id, an equals sign, and
the value of the field. Subsequent fields must be separated by an
ampersand ("&").

For example, to prefill the name field with "Derpy Hooves", use the URL:

        <MAIN_URL>/contact/#name=Derpy%20Hooves

To prefill the name as above and the subject with something else, use:

        <MAIN_URL>/contact/#name=Derpy%20Hooves&subject=something%20else

In the future, instead of making copies of the contact page all over the
site, which are hard to track and keep updated, simply point to the
contact page with a bit of prefilled fields.

To start this off, the header has been modified so that
"Resources -> Add a Resource" now points to the contact page with the
subject line prefilled as "Resource Suggestion for ACM Website". The
corresponding index.html page has been deleted. It can be brought back
from a prior commit if need be.

Also modified .gitignore to ignore temp emacs files