Fixed issues with Git repos on bottom of main page
Introduced a min-height requirement for the git repo boxes to make them
all even. These do *not* work when the screen is very very narrow, but
that is an edge case I'm willing to ignore for now. Otherwise, it
doesn't look bad at the other bootstrap size cases.

Added links to git repos for both the titles and the images. It was
previously just links on the images.

Also added an entry for the website repo, so now we look super
sexy with the scrolling repos!